requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
yt-dlp[default]>=2024.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.1.0,<4.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
aiofiles>=23.0.0
email-validator>=2.0.0
